Refactor SimpleEmail to be in new directory
